requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
Deprecated==1.2.13
idna==3.3
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.5.0
python-dotenv==0.19.2
requests==2.27.1
urllib3==1.26.8
wrapt==1.14.0
